VR component/requirements.txt
torch
torchvision
diffusers
transformers
accelerate
controlnet_aux
opencv-python
Pillow
numpy
timm
